requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0 # For Hidden Markov Models
optuna>=3.2.0

# Telegram parsing
telethon>=1.28.0

# Data processing
pandas>=2.0.0
numpy # Allow latest compatible version
python-dateutil>=2.8.2
nltk>=3.8.1

# Financial data and technical analysis
yfinance>=0.2.18
python-binance>=1.0.19 # For Binance API
pandas-ta>=0.3.14b # For custom strategy indicators
TA-Lib>=0.4.25  # Note: Requires separate C library installation (can be optional if pandas_ta is primary)
Backtesting>=0.3.3 # For strategy backtesting

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.2

# Async support
aiohttp>=3.8.4
asyncio-throttle>=1.0.2

# Additional utilities
python-dotenv>=1.0.0  # For environment variables
pyyaml>=6.0  # For config files
tqdm>=4.65.0  # Progress bars
colorlog>=6.7.0  # Colored logging

# Database (optional, for storing results)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6  # PostgreSQL adapter

# API framework (optional, for serving predictions)
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0

# Testing (optional)
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Linting and formatting (optional)
black>=23.3.0
flake8>=6.0.0
mypy>=1.2.0
python-telegram-bot
solana